+++ packages/common/src/core/schemas/getEntityEditorSchemas.ts
@@ -20,6 +20,10 @@ import {
   DiscussionEditorType,
   DiscussionEditorTypes,
 } from "../../discussions/_internal/DiscussionSchema";
+import {
+  PageEditorType,
+  PageEditorTypes,
+} from "../../pages/_internal/PageSchema";
 
 /**
  * defines the possible editor type values - these correspond
@@ -31,6 +35,7 @@ export const validEditorTypes = [
   ...InitiativeEditorTypes,
   ...SiteEditorTypes,
   ...DiscussionEditorTypes,
+  ...PageEditorTypes,
 ] as const;
 
 /**
@@ -99,6 +104,14 @@ export const getEntityEditorSchemas = async (
         "hub:discussion:create": () =>
           import("../../discussions/_internal/DiscussionUiSchemaCreate"),
       }[type as DiscussionEditorType]());
+      break;
+    case "page":
+      const { PageSchema } = await import("../../pages/_internal/PageSchema");
+      schema = cloneObject(PageSchema);
+
+      ({ uiSchema } = await {
+        "hub:page:edit": () => import("../../pages/_internal/PageUiSchemaEdit"),
+      }[type as PageEditorType]());
       break;
   }