+++ packages/common/src/core/types/IHubPage.ts
@@ -1,8 +1,11 @@
-import { IWithLayout } from "../traits/IWithLayout";
 import { IHubItemEntity } from "./IHubItemEntity";
-
+import { IWithLayout, IWithPermissions, IWithSlug } from "../traits";
 /**
  * DRAFT: Under development and more properties will likely be added
  * @internal
  */
-export interface IHubPage extends IHubItemEntity, IWithLayout {}
+export interface IHubPage
+  extends IHubItemEntity,
+    IWithLayout,
+    IWithPermissions,
+    IWithSlug {}